moved find_ methods to utils ...
