+++ threddsclient/nodes.py
@@ -1,24 +1,5 @@
-#!/usr/bin/env python
 """
 A Python view of a Thredds data server
-
-See thredds catalog description at: http://www.unidata.ucar.edu/software/thredds/current/tds/tutorial/CatalogPrimer.html
-
-author: Scott Wales <scott.wales@unimelb.edu.au>
-
-Copyright 2015 ARC Centre of Excellence for Climate Systems Science
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
 """
 
 from bs4 import BeautifulSoup as BSoup
@@ -93,9 +74,9 @@ class CollectionDataset(Dataset):
     def __init__(self, soup, catalog, skip):
         Dataset.__init__(self, soup, catalog)
         self.content_type = "application/directory"
-        from .catalog import find_datasets
+        from .utils import find_datasets
         self.datasets = find_datasets(soup, self.catalog, skip)
-        from .catalog import find_references
+        from .utils import find_references
         self.references = find_references(soup, self.catalog, skip)
 
     def is_collection(self):
